requirements.txt
instaloader==4.11
pillow==10.3.0